chore: MODE 관련 Channel 메소드 파일 분리
