+++ src/Channel.cpp
@@ -64,58 +64,3 @@ void Channel::displayTopic(Session *session) {
     }
     return NumericReply(RPL_TOPIC, this->topic) >> session;
 }
-
-void Channel::setInviteOnly(Channel::mode_t mode) {
-    inviteOnly = mode == ADD;
-}
-
-void Channel::setTopicRestricted(Channel::mode_t mode) {
-    topicRestricted = mode == ADD;
-}
-
-int Channel::setKey(Channel::mode_t mode, std::string const &key) {
-    if (key.empty()) {
-        return 0;
-    }
-    if (mode == ADD) {
-        this->key = key;
-        return 1;
-    }
-    this->key.clear();
-    return 0;
-}
-
-int Channel::setLimit(Channel::mode_t mode, std::string const &limit, Session *session) {
-    if (mode == REMOVE) {
-        this->limit = 0;
-        return 0;
-    }
-    if (limit.empty()) {
-        return 0;
-    }
-    try {
-        this->limit = Parser::parseLimit(limit);
-    } catch (std::exception &e) {
-        NumericReply(ERR_INVALIDMODEPARAM) << session << name << "+l" << limit >> session;
-    }
-    return 1;
-}
-
-int Channel::setOperator(Channel::mode_t mode, const std::string &nickname, Session *session) {
-    if (nickname.empty()) {
-        NumericReply(RPL_NAMREPLY, getOperatorList()) << session << PUBLIC_CHANNEL_SYMBOL << name >> session;
-        NumericReply(RPL_ENDOFNAMES) << session << name >> session;
-        return 0;
-    }
-    Session *participant = getParticipant(nickname);
-    if (!participant) {
-        NumericReply(ERR_USERNOTINCHANNEL) << session << nickname << name >> session;
-        return 1;
-    }
-    if (mode == ADD) {
-        operators.insert(participant);
-    } else {
-        operators.erase(participant);
-    }
-    return 1;
-}
